🍻 Main group 3 - customer-products

+++ back-end/src/database/migrations/20210825193958-create-product.js
@@ -14,7 +14,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.DOUBLE
+        type: Sequelize.DECIMAL(10,2)
       },
       url_image: {
         allowNull: false,
